Changed class names To reflect class stereotypes: entity, boundary and control
