--- src/User/Administrator.java
@@ -1,5 +1,0 @@
-package User;
-
-public class Administrator {
-    
-}
