+++ src/User/BloodType.java
@@ -1,9 +1,9 @@
-package Records;
+package User;
 public enum BloodType {
     A_NEGATIVE,
     A_POSITIVE,
     B_NEGATIVE,
     B_POSITIVE,
     O_NEGATIVE,
-    O_POSITIVE;
+    O_POSITIVE
 }
