+++ src/User/ContactInfo.java
@@ -1,4 +1,4 @@
-package Records;
+package User;
 public class ContactInfo {
     private int phoneNumber;
     private String email;
