+++ src/User/Doctor.java
@@ -1,93 +1,29 @@
 package User;
 
-import DatabaseManagers.DoctorRecordManager;
-import Records.DoctorRecord;
 
-import java.util.Scanner;
+import java.util.List;
 
-public class Doctor implements IHasMenu {
-    private DoctorRecord doctorRecord;
+public class Doctor {
+    private String doctorID;
+    private String name;
+    private List<Patient> patients;
+    //private List<Appointment> appointments;
 
-    public Doctor(String doctorID) throws Exception {
-        doctorRecord = DoctorRecordManager.loadDoctorRecord(doctorID);
+    public Doctor(String doctorID, String name, List<Patient> patients) {
+        this.doctorID = doctorID;
+        this.name = name;
+        this.patients = patients;
     }
 
-    String getDoctorID() {
-        return doctorRecord.getDoctorID();
+    public String getDoctorID() {
+        return doctorID;
     }
 
-    String getDoctorName() {
-        return doctorRecord.getName();
+    public String getName() {
+        return name;
     }
 
-    @Override
-    public void displayMenu() throws Exception {
-        while (true) {
-            System.out.println("Welcome Doctor " + doctorRecord.getName() + ",\n");
-            System.out.println("(1) View Patient Medical Records\n" + "(2) Update Patient Medical Records (Coming Soon)\n" + "(3) Appointments (Coming Soon)\n" + "(4) Log Out\n");
-            Scanner sc = new Scanner(System.in);
-            int choice;
-            try {
-                choice = Integer.parseInt(sc.nextLine());
-            } catch (NumberFormatException e) {
-                choice = 0;
-            }
-            System.out.println();
-            switch (choice) {
-                case 1:
-                    viewPatientRecords();
-                    break;
-                case 2:
-                    break;
-                case 3:
-                    break;
-                case 4:
-                    System.out.println("Logging out...");
-                    System.out.println("Successfully logged out!\n\n");
-                    return;
-                default:
-                    System.out.println("Invalid choice\n");
-                    break;
-            }
-        }
-    }
-
-
-    private void viewPatientRecords() throws Exception {
-        Scanner sc = new Scanner(System.in);
-        System.out.println("<<Patient Records View>>");
-        System.out.println("Enter (X) to return\n");
-        System.out.println("Patient List:");
-        for (Patient p : doctorRecord.getPatients()) {
-            System.out.println(p);
-        }
-        System.out.print("Enter Patient ID to view record: ");
-        String patientID = sc.nextLine().toUpperCase();
-        if (patientID.equals("X")) {
-            return;
-        }
-        System.out.println();
-        Patient patient = new Patient(patientID);
-        if (patient.getPatientRecord() == null) {
-            System.out.println("Invalid Patient ID. Please try again.\n");
-            viewPatientRecords();
-            return;
-        }
-        boolean authorized = false;
-        for (Patient p : doctorRecord.getPatients()) {
-            if (p.getPatientID().equals(patient.getPatientID())) {
-                authorized = true;
-                break;
-            }
-        }
-
-        if (!authorized) {
-            System.out.println("You are not authorized to view this patient's record.\n");
-            viewPatientRecords();
-            return;
-        }
-        System.out.println(patient.getPatientRecord() + "\n");
-        System.out.println("Press enter to continue...\n");
-        sc.nextLine();
+    public List<Patient> getPatients() {
+        return patients;
     }
 }
