--- src/User/Pharmacist.java
@@ -1,5 +1,0 @@
-package User;
-
-public class Pharmacist {
-    
-}
