+++ src/UserMenus/AdministratorMenu.java
@@ -1,0 +1,5 @@
+package UserMenus;
+
+public class AdministratorMenu {
+    
+}
