+++ src/UserMenus/IHasMenu.java
@@ -1,4 +1,4 @@
-package User;
+package UserMenus;
 
 public interface IHasMenu {
     public void displayMenu() throws Exception;
