+++ src/UserMenus/PharmacistMenu.java
@@ -1,0 +1,5 @@
+package UserMenus;
+
+public class PharmacistMenu {
+    
+}
